Refactored my code to modify the company holidays list to a component + Added the default option for longweekend and excluded mixItUp for now
